+++ mango_tango_cib/components/splash.py
@@ -1,5 +1,5 @@
 from mango_tango_cib.terminal_tools import clear_terminal, wait_for_key
-from meta import get_version
+from mango_tango_cib.meta import get_version
 
 
 def splash():
